+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - NPC don't drop weapon until HP level above 5, then NPC start panic as before #
 - Add simple mini-map #10
 - Add "change name" button #14
+- Reset frags on death, add scores #18
+- Slower NPC spawning
+- Add coin and diamond items for scores
+- Add shield item and health-kit to add +2 hp
+- Add online users info in the main menu
+- Add time to reload weapon mag
 
 ## 1.0.1
 - NPC are also fighting